SDL_Window & GameLoop Class Created
An SDL window is now created using pre-defined width of the window and calculating the height using the aspect ratio (16:9).

Code for initialising and destroying this window (and SDL subsystems) has been added to functions in a GameLoop class. These functions are called from main().

A "CommonIncludes" file has been created which will be used in each file and stores commonly used header files, making it easier to reference other files where required.